Refine dashboard profile provider structure
